requirements.txt
-e git+https://github.com/chwe17/pyGPIO.git#egg=pyGPIO-0.2.1
python-dotenv
